Moved departure to separate folder
